Generalize intrinsic invocation dispatch
Copilot-Session: 946bd35e-0c84-49f0-90ff-a15d522d34f8
